Making backup-create task work on localhost.

+++ contrib/backup-create.sh
@@ -15,37 +15,44 @@
 # be a dedicated backup environment.
 
 source $DETA/backup.sh
-source $DETA/transfer.sh
-source $DETA/invoke.sh
 
 role SOURCE # The env to backup from.
 role TARGET # The env to store the backup in.
 
 dry
 
-TMP=$(mktemp -d -t deta)
-defer rm -rf $TMP
+if [[ $SOURCE_HOST != 'localhost' ]]; then
+	source $DETA/transfer.sh
+	source $DETA/invoke.sh
 
-PS3="Backup to base onto (most recent first): "
-select file in $(ls -t $TARGET_PATH/*) "none"; do
-	if [[ $file == "none" ]]; then
+	TMP=$(mktemp -d -t deta)
+	defer rm -rf $TMP
+
+	PS3="Select a previous backup to base this one onto (most recent first): "
+	select file in $(ls -t $TARGET_PATH/*) "none"; do
+		if [[ $file == "none" ]]; then
+			break
+		fi
+		dearchive $file $TMP
 		break
-	fi
-	dearchive $file $TMP
-	break
-done
+	done
 
-# Uncomment and modify to run tasks on source.
-# run_ssh $SOURCE_USER@$SOURCE_HOST <<-SESSION
-# 	cd $SOURCE_PATH/scripts
-# 	./deta db/dump.sh
-# SESSION
+	# Uncomment and modify to run tasks on source.
+	# run_ssh $SOURCE_HOST <<-SESSION
+	#  cd $SOURCE_PATH/bin
+	#  ./deta.sh -c ../config/deta db/dump.sh
+	# SESSION
 
-sync_sanity "$SOURCE_USER@$SOURCE_HOST:$SOURCE_PATH/*" $TMP "$SOURCE_IGNORE"
+	sync_sanity "$SOURCE_USER@$SOURCE_HOST:$SOURCE_PATH/*" $TMP "$BACKUP_IGNORE"
+	set +o errexit
+	sync "$SOURCE_USER@$SOURCE_HOST:$SOURCE_PATH/*" $TMP "$BACKUP_IGNORE"
+	set -o errexit
 
-# Temporarily disable errexit (Files not synced are considered a non critical error).
-set +o errexit
-sync "$SOURCE_USER@$SOURCE_HOST:$SOURCE_PATH/*" $TMP "$SOURCE_IGNORE"
-set -o errexit
+	archive $TMP $TARGET_PATH ${SOURCE_HOST}-${SOURCE_NAME}
+else
+	# Uncomment and modify to run tasks on source.
+	# cd $SOURCE_PATH/bin
+	# ./deta.sh -c ../config/deta db/dump.sh
 
-archive $TMP $TARGET_PATH ${SOURCE_HOST}-${SOURCE_NAME}
+	archive $SOURCE_PATH $TARGET_PATH ${SOURCE_HOST}-${SOURCE_NAME}
+fi
